Move form builder to its own package
