+++ packages/form-builder/package.json
@@ -1,0 +1,18 @@
+{
+	"author": "Event Espresso",
+	"name": "@eventespresso/form-builder",
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/eventespresso/barista.git"
+	},
+	"private": true,
+	"version": "1.0.0",
+	"main": "src/index.ts",
+	"scripts": {
+		"lint": "eslint ./src/**/*.{ts,tsx} --format=codeframe",
+		"update-deps": "npx ncu -p yarn -u",
+		"lint-fix": "eslint ./src/**/*.{ts,tsx} --fix"
+	},
+	"sideEffects": false,
+	"license": "GPL-3.0"
+}
