+++ packages/form-builder/src/FormElement/Tabs/Styles.tsx
@@ -1,7 +1,6 @@
 import { __ } from '@eventespresso/i18n';
+import { TextInputWithLabel, TextareaWithLabel } from '@eventespresso/ui-components';
 
-import { TextInputWithLabel } from '../../../text-input';
-import { TextareaWithLabel } from '../../../Textarea';
 import { useUpdateElement } from '../useUpdateElement';
 
 import type { FormElementProps } from '../../types';
