+++ packages/ui-components/src/index.ts
@@ -30,7 +30,6 @@ export * from './EntityList';
 export * from './EntityPaperFrame';
 export * from './EspressoTable';
 export * from './FilterBarFilter';
-export * from './FormBuilder';
 export * from './Grid';
 export * from './Heading';
 export * from './HelpIcon';
@@ -38,6 +37,7 @@ export * from './InlineEdit';
 export * from './InlineNotification';
 export * from './ItemCount';
 export * from './Legend';
+export * from './Label';
 export * from './LoadingNotice';
 export * from './MediumCalendarDate';
 export * from './Modal';
